feat(#3443): Move data explorer dashboard to dashboard UI module
