+++ ui/src/app/assets/dialog/manage-asset-links/manage-asset-links-dialog.component.html
@@ -290,12 +290,12 @@
                             selectLink(
                                 $event.checked,
                                 element.elementId,
-                                element.name,
+                                element.baseAppearanceConfig.widgetTitle,
                                 'data-view'
                             )
                         "
                     >
-                        {{ element.name }}
+                        {{ element.baseAppearanceConfig.widgetTitle }}
                     </mat-checkbox>
                 </div>
             </div>
