+++ ui/src/app/dashboard/components/overview/dashboard-overview.component.html
@@ -22,5 +22,31 @@
     [resourceCount]="0"
     (filterIdsEmitter)="applyDashboardFilters($event)"
 >
-    <sp-basic-view [showBackLink]="false" [padding]="true"> </sp-basic-view>
+    <sp-basic-view [showBackLink]="false" [padding]="true">
+        <div
+            nav
+            fxFlex="100"
+            fxLayoutAlign="start center"
+            fxLayout="row"
+            class="pl-10"
+        >
+            <button
+                mat-button
+                mat-raised-button
+                color="accent"
+                data-cy="open-new-dashboard-dialog"
+                (click)="openNewDashboardDialog()"
+                class="mr-10"
+                *ngIf="hasDashboardWritePrivileges"
+            >
+                <i class="material-icons">add</i>
+                <span>New Dashboard</span>
+            </button>
+        </div>
+        <div fxFlex="100" fxLayout="column">
+            <sp-data-explorer-dashboard-overview
+                (resourceCountEmitter)="resourceCount = $event"
+            ></sp-data-explorer-dashboard-overview>
+        </div>
+    </sp-basic-view>
 </sp-asset-browser>
