+++ ui/src/app/dashboard/components/panel/dashboard-toolbar/dashboard-toolbar.component.scss
