+++ ui/src/app/dashboard/components/panel/dashboard-widget-selection-panel/dashboard-widget-selection-panel.component.scss
