+++ ui/src/app/dashboard/components/panel/dashboard-widget-selection-panel/data-view-selection/data-view-selection.component.html
