+++ ui/src/app/dashboard/components/panel/data-explorer-dashboard-panel.component.html
@@ -17,7 +17,7 @@
   -->
 
 <sp-basic-view
-    [backLinkTarget]="['dataexplorer']"
+    [backLinkTarget]="['dashboard']"
     [showBackLink]="true"
     *ngIf="dashboardLoaded"
 >
