+++ ui/src/app/dashboard/components/panel/data-explorer-dashboard-panel.component.scss
