+++ ui/src/app/dashboard/components/widget-view/abstract-widget-view.directive.ts
