+++ ui/src/app/dashboard/components/widget-view/grid-view/data-explorer-dashboard-grid.component.scss
