+++ ui/src/app/dashboard/components/widget-view/slide-view/data-explorer-dashboard-slide-view.component.html
