+++ ui/src/app/dashboard/dashboard.module.ts
@@ -59,6 +59,17 @@ import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatButtonToggleModule } from '@angular/material/button-toggle';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatSliderModule } from '@angular/material/slider';
+import { DataExplorerSharedModule } from '../data-explorer-shared/data-explorer-shared.module';
+import { DataExplorerDashboardPanelComponent } from './components/panel/data-explorer-dashboard-panel.component';
+import { DataExplorerPanelCanDeactivateGuard } from '../data-explorer-shared/services/data-explorer-panel.can-deactivate.guard';
+import { DataExplorerDashboardGridComponent } from './components/widget-view/grid-view/data-explorer-dashboard-grid.component';
+import { DataExplorerDashboardSlideViewComponent } from './components/widget-view/slide-view/data-explorer-dashboard-slide-view.component';
+import { DataExplorerDashboardToolbarComponent } from './components/panel/dashboard-toolbar/dashboard-toolbar.component';
+import { DataExplorerDashboardWidgetSelectionPanelComponent } from './components/panel/dashboard-widget-selection-panel/dashboard-widget-selection-panel.component';
+import { DataExplorerDataViewPreviewComponent } from './components/panel/dashboard-widget-selection-panel/data-view-selection/data-view-preview/data-view-preview.component';
+import { DataExplorerDataViewSelectionComponent } from './components/panel/dashboard-widget-selection-panel/data-view-selection/data-view-selection.component';
+import { DataExplorerEditDashboardDialogComponent } from './dialogs/edit-dashboard/data-explorer-edit-dashboard-dialog.component';
+import { SpDataExplorerDashboardOverviewComponent } from './components/overview/data-explorer-dashboard-overview/data-explorer-dashboard-overview.component';
 
 @NgModule({
     imports: [
@@ -102,6 +113,7 @@ import { MatSliderModule } from '@angular/material/slider';
         PlatformServicesModule,
         ServicesModule,
         SharedUiModule,
+        DataExplorerSharedModule,
         RouterModule.forChild([
             {
                 path: '',
@@ -110,11 +122,32 @@ import { MatSliderModule } from '@angular/material/slider';
                         path: '',
                         component: DashboardOverviewComponent,
                     },
+                    {
+                        path: ':id',
+                        component: DataExplorerDashboardPanelComponent,
+                        canDeactivate: [DataExplorerPanelCanDeactivateGuard],
+                    },
+                    {
+                        path: ':id/:startTime/:endTime',
+                        component: DataExplorerDashboardPanelComponent,
+                        canDeactivate: [DataExplorerPanelCanDeactivateGuard],
+                    },
                 ],
             },
         ]),
     ],
-    declarations: [DashboardOverviewComponent],
+    declarations: [
+        DashboardOverviewComponent,
+        DataExplorerDashboardGridComponent,
+        DataExplorerDashboardPanelComponent,
+        DataExplorerDashboardSlideViewComponent,
+        DataExplorerDashboardToolbarComponent,
+        DataExplorerDashboardWidgetSelectionPanelComponent,
+        DataExplorerDataViewPreviewComponent,
+        DataExplorerDataViewSelectionComponent,
+        DataExplorerEditDashboardDialogComponent,
+        SpDataExplorerDashboardOverviewComponent,
+    ],
     providers: [],
     exports: [],
 })
