+++ ui/src/app/dashboard/dialogs/edit-dashboard/data-explorer-edit-dashboard-dialog.component.html
