+++ ui/src/app/data-explorer/components/overview/data-explorer-overview.component.ts
@@ -16,8 +16,7 @@
  *
  */
 
-import { Component, ViewChild } from '@angular/core';
-import { Dashboard } from '@streampipes/platform-services';
+import { Component } from '@angular/core';
 import {
     CurrentUserService,
     DialogService,
@@ -25,10 +24,8 @@ import {
 } from '@streampipes/shared-ui';
 import { AuthService } from '../../../services/auth.service';
 import { SpDataExplorerRoutes } from '../../data-explorer.routes';
-import { SpDataExplorerDashboardOverviewComponent } from './data-explorer-dashboard-overview/data-explorer-dashboard-overview.component';
 import { SpDataExplorerOverviewDirective } from './data-explorer-overview.directive';
-import { DataExplorerRoutingService } from '../../services/data-explorer-routing.service';
-import { DataExplorerDashboardService } from '../../services/dashboard.service';
+import { DataExplorerRoutingService } from '../../../data-explorer-shared/services/data-explorer-routing.service';
 
 @Component({
     selector: 'sp-data-explorer-overview',
@@ -36,15 +33,11 @@ import { DataExplorerDashboardService } from '../../services/dashboard.service';
     styleUrls: ['./data-explorer-overview.component.scss'],
 })
 export class DataExplorerOverviewComponent extends SpDataExplorerOverviewDirective {
-    @ViewChild(SpDataExplorerDashboardOverviewComponent)
-    dashboardOverview: SpDataExplorerDashboardOverviewComponent;
-
     resourceCount = 0;
 
     constructor(
         public dialogService: DialogService,
         private breadcrumbService: SpBreadcrumbService,
-        private dataExplorerDashboardService: DataExplorerDashboardService,
         authService: AuthService,
         currentUserService: CurrentUserService,
         routingService: DataExplorerRoutingService,
@@ -58,38 +51,11 @@ export class DataExplorerOverviewComponent extends SpDataExplorerOverviewDirecti
         );
     }
 
-    openNewDashboardDialog() {
-        const dataViewDashboard: Dashboard = {
-            dashboardGeneralSettings: {},
-            widgets: [],
-            name: '',
-            dashboardLiveSettings: {
-                refreshModeActive: false,
-                refreshIntervalInSeconds: 10,
-                label: 'Off',
-            },
-        };
-
-        this.openDashboardModificationDialog(true, dataViewDashboard);
-    }
-
     createNewDataView(): void {
         this.routingService.navigateToDataView(true);
     }
 
-    openDashboardModificationDialog(createMode: boolean, dashboard: Dashboard) {
-        const dialogRef =
-            this.dataExplorerDashboardService.openDashboardModificationDialog(
-                createMode,
-                dashboard,
-            );
-
-        dialogRef.afterClosed().subscribe(() => {
-            this.dashboardOverview.getDashboards();
-        });
-    }
-
     applyDashboardFilters(elementIds: Set<string> = new Set<string>()): void {
-        this.dashboardOverview.applyDashboardFilters(elementIds);
+        //this.da.applyDashboardFilters(elementIds);
     }
 }
